requirements.txt
# Core dependencies
fasttext>=0.9.2
openai>=1.3.0
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-pinecone>=0.1.0
pinecone-client>=2.2.4
python-dotenv>=1.0.0
streamlit>=1.29.0

# API dependencies
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# Document processing
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.2
nltk>=3.8.1

# Monitoring and visualization
rich>=13.5.0
tqdm>=4.66.1

# Testing
pytest>=7.4.0

# Utilities
argparse>=1.4.0
requests>=2.31.0 